fixed bugs and added new router

+++ app/routers/banner/model/bannerDB.statics.js
@@ -33,7 +33,11 @@ const createImageById = function (bannerId, image) {
             return Banner.findOneAndUpdate({ _id: bannerId }, { $set: { image: path } })
                 .then((result) => {
                     if (result.image) {
-                        fs.unlinkSync(PATH.join(__dirname, '..', '..','static', 'imgs', result.image));
+                        fs.stat(PATH.join(__dirname, '..', '..','static', 'imgs', result.image), function(err, stat) {
+                            if(err == null){
+                                fs.unlinkSync(PATH.join(__dirname, '..', '..','static', 'imgs', result.image));
+                            }
+                         }); 
                     } else {
                         console.log("this is 1");
                     }
