+++ app/routers/service/model/serviceDB.statics.js
@@ -31,7 +31,11 @@ const createImageById = function (serviceId, image) {
             return Service.findOneAndUpdate({ _id: serviceId }, { $set: { image: path } })
                 .then((result) => {
                     if (result.image) {
-                        fs.unlinkSync(PATH.join(__dirname, '..', '..','static', 'imgs', result.image));
+                        fs.stat(PATH.join(__dirname, '..', '..','static', 'imgs', result.image), function(err, stat) {
+                            if(err == null){
+                                fs.unlinkSync(PATH.join(__dirname, '..', '..','static', 'imgs', result.image));
+                            }
+                         }); 
                     } else {
                         console.log("this is 1");
                     }
