+++ app/routers/staff/model/staffDB.static.js
@@ -49,7 +49,11 @@ const createImageById = function (staffId, image) {
             return Staff.findOneAndUpdate({ _id: staffId }, { $set: { image: path } })
                 .then((result) => {
                     if (result.image) {
-                        fs.unlinkSync(PATH.join(__dirname, '..', '..', 'static', 'imgs', result.image));
+                        fs.stat(PATH.join(__dirname, '..', '..','static', 'imgs', result.image), function(err, stat) {
+                            if(err == null){
+                                fs.unlinkSync(PATH.join(__dirname, '..', '..','static', 'imgs', result.image));
+                            }
+                         }); 
                     } else {
                         console.log("this is 1");
                     }
